Hotfixes: update user with img with existing email

+++ src/main/java/com/dam/kanpeki/controller/UserController.java
@@ -223,12 +223,22 @@ public class UserController {
 
 		if (opUser.isPresent()) {
 
-			if (file != null) {
+			ResponseUserDTO responseUserDTO = uService.updateUser(u, file, id);
+
+			if (responseUserDTO != null && file != null) {
 				// Eliminamos la imagen anterior del almacenamiento
 				storeService.delete(opUser.get().getUrlImage());
 			}
 
-			return ResponseEntity.ok(uService.updateUser(u, file, id));
+			return ResponseEntity.ok(responseUserDTO);
+
+//			if (file != null) {
+//				
+//				// Eliminamos la imagen anterior del almacenamiento
+//				storeService.delete(opUser.get().getUrlImage());
+//			}
+//
+//			return ResponseEntity.ok(uService.updateUser(u, file, id));
 
 		} else {
 			throw new DataNotFoundException(KanpekiConstants.EMPTY_STRING);
